api/requirements.txt
boto3==1.13.26
Django==3.0.6
django-storages==1.9.1
dj-database-url==0.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.0
gunicorn==20.0.4
pillow==2.1.0
psycopg2-binary==2.8.5
